# src/ActionLedger.Web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ActionLedger API",
    "description": "Turns meeting notes into tracked, attributable action items. Every route lives under /api/v1 except /health, /openapi, and /swagger. Errors are RFC 9457 ProblemDetails whose type is one of validation, unauthorized, forbidden, not-found, or conflict.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/meetings": {
      "post": {
        "tags": [
          "Meetings"
        ],
        "summary": "Creates a meeting.",
        "description": "The meeting is attributed to the caller's own token: createdByUserId comes from the sub claim and never from the request body. A title and date that match an existing meeting are a conflict.",
        "operationId": "CreateMeeting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeetingCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeetingCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeetingCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingCreatedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingCreatedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingCreatedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Meetings"
        ],
        "summary": "Lists meetings, newest first.",
        "description": "Ordered by meeting date descending, then creation time, then id, and paged by the shared page/pageSize window. runCount and trackedActionCount are published now and are 0 until extraction runs and tracked actions exist.",
        "operationId": "ListMeetings",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/pageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMeetingSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMeetingSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMeetingSummaryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/meetings/{id}/notes": {
      "put": {
        "tags": [
          "Meetings"
        ],
        "summary": "Attaches the meeting's notes. Once.",
        "description": "Notes are stored byte-for-byte with a SHA-256 of exactly those bytes. They cannot be changed after saving: a second call is 409, and no endpoint updates or deletes them.",
        "operationId": "SaveMeetingNotes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveMeetingNotesCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveMeetingNotesCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveMeetingNotesCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingNotesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingNotesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingNotesDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/meetings/{id}/runs": {
      "post": {
        "tags": [
          "Meetings"
        ],
        "summary": "Starts an extraction run against this meeting's notes.",
        "description": "The run is synchronous and takes no body; the actor comes from the sub claim. A failed extraction is still 201 with outcome Failed and a reason, because a model that returned nonsense is not an HTTP error. A meeting may have any number of runs.",
        "operationId": "StartExtractionRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Meetings"
        ],
        "summary": "Lists this meeting's extraction runs, newest first.",
        "description": "Ordered by start time descending, then id. Each run carries provider, model, prompt version, outcome and failure reason, plus proposalCount and pendingCount computed server-side. A meeting with no runs answers an empty array.",
        "operationId": "ListExtractionRuns",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/meetings/{id}": {
      "get": {
        "tags": [
          "Meetings"
        ],
        "summary": "Reads one meeting, with its notes.",
        "description": "notes is null while none have been attached, and otherwise carries the text exactly as it was pasted together with the SHA-256 of those bytes.",
        "operationId": "GetMeeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/runs/{id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Reads one extraction run, with its metadata and proposals.",
        "description": "Carries provider, model, prompt version, schema version, start time, duration, both token counts, outcome, failure reason and warnings. Proposals come back in AI order with isLowConfidence and suggestedOwnerUserId already computed server-side.",
        "operationId": "GetExtractionRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMeetingCommand": {
        "required": [
          "title",
          "meetingDate"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "meetingDate": {
            "type": "string",
            "format": "date"
          },
          "attendees": {
            "type": "array",
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            }
          }
        }
      },
...
      "MeetingCreatedDto": {
        "required": [
          "id",
          "createdByUserId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdByUserId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveMeetingNotesCommand": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "maxLength": 50000,
            "minLength": 1,
            "type": "string"
          }
        }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "description": "The 1-based page to return.",
        "schema": {
          "minimum": 1,
          "type": "integer",
          "format": "int32",
          "default": 1
        }
      },
      "pageSize": {
        "name": "pageSize",
        "in": "query",
        "description": "Items per page. Values above 200 are clamped to 200.",
        "schema": {
          "maximum": 200,
          "minimum": 1,
          "type": "integer",
          "format": "int32",
          "default": 50
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "An HS256 token from POST /api/v1/auth/login carrying sub, name, and role. Send it as: Authorization: Bearer {token}.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}